# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface InferenceHealth {
  status: string;
  inference: string;
  model_configured: boolean;
  model_name: string;
}

export interface DetectionRequest {
  /** Base64-encoded image or data URL containing a real CCTV frame */
  image_base64: string;
  /** @minimum 1 */
  capacity?: number;
  /**
     * @minimum 0.01
     * @maximum 0.99
     */
  confidence?: number;
}

export interface DetectionBox {
  x1: number;
  y1: number;
  x2: number;
  y2: number;
  confidence: number;
}

export type DetectionResponseCrowdLevel = typeof DetectionResponseCrowdLevel[keyof typeof DetectionResponseCrowdLevel];


export const DetectionResponseCrowdLevel = {
  Low: 'Low',
  Medium: 'Medium',
  High: 'High',
} as const;

export interface DetectionResponse {
  ok: boolean;
  count: number;
  boxes: DetectionBox[];
  crowd_level: DetectionResponseCrowdLevel;
  /** @nullable */
  crowd_percentage: number | null;
  model: string;
}

export type GeminiChatRequestMessagesItemRole = typeof GeminiChatRequestMessagesItemRole[keyof typeof GeminiChatRequestMessagesItemRole];


export const GeminiChatRequestMessagesItemRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export type GeminiChatRequestMessagesItem = {
  role: GeminiChatRequestMessagesItemRole;
  /** @maxLength 4000 */
  content: string;
};

export type GeminiChatRequestContext = { [key: string]: unknown };

export interface GeminiChatRequest {
  /** @maxItems 16 */
  messages: GeminiChatRequestMessagesItem[];
  context?: GeminiChatRequestContext;
}

export interface GeminiChatResponse {
  message: string;
  assistant: string;
}