# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gemini/chat:
    post:
      operationId: chatWithMahaFlowAI
      tags: [gemini]
      summary: Ask MahaFlow AI a transportation question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeminiChatRequest"
      responses:
        "200":
          description: MahaFlow AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiChatResponse"
        "400":
          description: Invalid request
        "429":
          description: Rate limited
        "502":
          description: AI provider unavailable
components:
  schemas:
...
    GeminiChatRequest:
      type: object
      properties:
        messages:
          type: array
          maxItems: 16
          items:
            type: object
            properties:
              role:
                type: string
                enum: [user, assistant]
              content:
                type: string
                maxLength: 4000
            required: [role, content]
        context:
          type: object
          additionalProperties: true
      required: [messages]
    GeminiChatResponse:
      type: object
      properties:
        message:
          type: string
        assistant:
          type: string
      required: [message, assistant]