# lib/api-zod/src/generated/types/geminiChatRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GeminiChatRequestContext } from './geminiChatRequestContext';
import type { GeminiChatRequestMessagesItem } from './geminiChatRequestMessagesItem';

export interface GeminiChatRequest {
  /** @maxItems 16 */
  messages: GeminiChatRequestMessagesItem[];
  context?: GeminiChatRequestContext;
}